pyproject.toml
[project]
name = "stock-selection"
version = "0.3.0"
description = "Opinionated foundation for a sector-relative stock selection application"
readme = "README.md"
requires-python = ">=3.12,<3.14"
authors = [
  { name = "Prashant Wani" }
]
dependencies = [
  "numpy>=2.1.3",
  "pandas>=2.2.3",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.8.1",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.6.1",
  "scipy>=1.14.1",
  "typer>=0.15.1",
  "rich>=13.9.4"
]

[project.scripts]
stock-selection = "stock_selection.cli.main:app"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.6",
  "pyright>=1.1.394",
  "pre-commit>=4.1.0"
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_selection"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.coverage.run]
source = ["src/stock_selection"]
